AncienneMethode/requirements.txt
# Core dependencies
streamlit>=1.28.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
matplotlib>=3.5.0
torch>=1.11.0
torchvision>=0.12.0

# SAM dependencies
segment-anything>=1.0

# MiDaS dependencies
timm>=0.6.0

# Stable Diffusion (optional but recommended)
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.12.0
xformers>=0.0.16  # Optional, for memory optimization

# Additional utilities
scipy>=1.7.0
scikit-image>=0.18.0
tqdm>=4.62.0

# For CUDA support (uncomment if using GPU)
# torch>=1.11.0+cu113
# torchvision>=0.12.0+cu113
# --extra-index-url https://download.pytorch.org/whl/cu113